refactor out push logic to pusher
